Use a frame to move items on each UI update
Instead of trying to move all items within a single UI update. This is
more flexible because it allows for the possibility of performing a
compaction operation across multiple UI updates. This will be necessary
in a situation where the addon needs to move an item into a slot whose
item the addon just moved out, but which the UI has not yet finished
moving. The frame is shown only upon starting a compaction operation and
hidden when complete, so the OnUpdate handler is only called when
needed.