Simplify move selection
Movement of each cursor to the next slot, checking the eligibility of
each slot as a source/destination, and validity of a particular
potential move, are now separate. Selection of the To slot now occurs
first because the criteria for the To slot are always simpler. If the To
cursor encounters an empty slot, then either that slot is the next
destination slot or the compaction operation is complete. A slot is now
never set to a number outside the bounds of the current bag, whereas
before it was always initially set just outside the bounds, which was
confusing.