Support iterative strategies
Iterative strategies will have their From slot return to the same
position as the To slot, with the intent of again looping through all
remaining slots, after each move. As this implies, each subsequent move
may depend on the previous, so after each move Borganizer waits for the
locks on moved items to release before moving the cursors.